correção de sintaxe

+++ idades.c
@@ -20,7 +20,7 @@ int main() {
     scanf("%d", &idade2);
     getchar();
 
-    media = (idade1 + idade2) / 2.0; 
+    media = (double)(idade1 + idade2) / 2; 
     printf("A idade media de %s e %s eh de %.1lf anos\n\n", nome1, nome2, media);    
 
     return 0;
